Two issue dealing with failure to load
Fix bug (thanks @salandora) where bad url or down server will rapidly
generate requests
I was overloading the OctoPrint snapshot setting. That one is for
OctoPrint to request for timelapse. It's possible that users will want
the same one, but they may not. In particular, a relative one is
desirable for snapstream where an absolute url is best for the
timelapse.

Considered using the tunnel /download/snapshot/current as the default,
but it increases the load on the octoprint server more than the direct
url. Users can still set it to that if it works better for their proxy
scheme.